+++ COFantasy.js
@@ -3189,6 +3189,27 @@ var COFantasy = COFantasy || function() {
     return bonus;
   }
 
+  var competenceNameRegExp = new RegExp(/^(repeating_competences_.*_)comp_nom$/);
+  var competenceNameCapaRegExp = new RegExp(/^(repeating_competences_.*_)comp_nomCapa$/);
+  var competenceTotalRegExp = new RegExp(/^(repeating_competences_.*_)comp_bonusTotal$/);
+  var competenceCaracsRegExp = new RegExp(/^(repeating_competences_.*_)comp_caracs$/);
+  var competenceCaracRegExp = new RegExp(/^(repeating_competences_.*_)comp_carac$/);
+
+  function validateCompetence(carac, comp, prefix, compInfos, prefixesNom) {
+    if (compInfos.total === undefined) return;
+    var res = {total: compInfos.total, bestFit:false};
+    if (compInfos.carac == carac) res.bestFit = true;
+    if (!res.bestFit && compInfos.caracs === undefined) return;
+    if (compInfos.caracs && compInfos.caracs.length > 0) {
+      if (compInfos.caracs.includes(carac)) res.bestFit = true;
+      else if (compInfos.carac === undefined) {
+        prefixesNom.delete(prefix);
+        return;
+      }
+    }
+    return res;
+  }
+
   //retourne un entier
   // evt n'est défini que si la caractéristique est effectivement utlilisée
   function bonusTestCarac(carac, personnage, options, testId, evt, explications) {
@@ -3207,15 +3228,6 @@ var COFantasy = COFantasy || function() {
     var bonusAspectDuDemon;
     switch (carac) {
       case 'DEX':
-        var malusArmure = 0;
-        if (ficheAttributeAsInt(personnage, 'DEFARMUREON', 1))
-          malusArmure += ficheAttributeAsInt(personnage, 'DEFARMUREMALUS', 0);
-        if (ficheAttributeAsInt(personnage, 'DEFBOUCLIERON', 1))
-          malusArmure += ficheAttributeAsInt(personnage, 'DEFBOUCLIERMALUS', 0);
-        if (malusArmure > 0) {
-          expliquer("Armure : -" + malusArmure + " au jet de DEX");
-          bonus -= malusArmure;
-        }
         if (attributeAsBool(personnage, 'agrandissement')) {
           expliquer("Agrandi : -2 au jet de DEX");
           bonus -= 2;
@@ -3267,6 +3279,134 @@ var COFantasy = COFantasy || function() {
           bonus += bonusAspectDuDemon;
         }
         break;
+    }
+    var bonusCompetence;
+    if (options && options.competence) {
+      var comp = options.competence.trim().toLowerCase();
+      var attributes = findObjs({
+        _type: 'attribute',
+        _characterid: personnage.charId,
+      });
+      var competences = {};
+      var prefixesNom = new Set();
+      var res;
+      attributes.find(function(attr) {
+        var attrName = attr.get('name');
+        var m = competenceNameRegExp.exec(attrName);
+        if (!m) m = competenceNameCapaRegExp.exec(attrName);
+        if (m) {
+          if (attr.get('current').trim().toLowerCase() != comp) return false;
+          prefixesNom.add(m[1]);
+          var compInfos = competences[m[1]];
+          if (compInfos === undefined) return false;
+          res = validateCompetence(carac, comp, m[1], compInfos, prefixesNom);
+          if (res === undefined) return false;
+          bonusCompetence = res.total;
+          return res.bestFit;
+        }
+        m = competenceCaracRegExp.exec(attrName);
+        if (m) {
+          competences[m[1]] = competences[m[1]] || {};
+          competences[m[1]].carac = attr.get('current');
+          if (!prefixesNom.has(m[1])) return false;
+          res = validateCompetence(carac, comp, m[1], competences[m[1]], prefixesNom);
+          if (res === undefined) return false;
+          bonusCompetence = res.total;
+          return res.bestFit;
+        }
+        m = competenceCaracsRegExp.exec(attrName);
+        if (m) {
+          competences[m[1]] = competences[m[1]] || {};
+          competences[m[1]].caracsLimitees = attr.get('current').split(',');
+          if (!prefixesNom.has(m[1])) return false;
+          res = validateCompetence(carac, comp, m[1], competences[m[1]], prefixesNom);
+          if (res === undefined) return false;
+          bonusCompetence = res.total;
+          return res.bestFit;
+        }
+        m = competenceTotalRegExp.exec(attrName);
+        if (!m) return false;
+        competences[m[1]] = competences[m[1]] || {};
+        competences[m[1]].total = parseInt(attr.get('current'));
+        if (isNaN(competences[m[1]].total)) competences[m[1]] = 0;
+        if (!prefixesNom.has(m[1])) return false;
+        res = validateCompetence(carac, comp, m[1], competences[m[1]], prefixesNom);
+        if (res === undefined) return false;
+        bonusCompetence = res.total;
+        return res.bestFit;
+      });
+      if (bonusCompetence === undefined) {
+        var found;
+        prefixesNom.forEach(function(prefix) {
+          if (found) return;
+          var compInfos = competences[prefix];
+          if (compInfos === undefined) 
+            compInfos = {total:0, carac:'FOR', caracs:[]};
+          else {
+            if (compInfos.total === undefined) compInfos.total = 0;
+            if (compInfos.carac === undefined) compInfos.carac = 'FOR';
+            if (compInfos.caracs === undefined) compInfos.caracs = [];
+          }
+          var res = validateCompetence(carac, comp, prefix, compInfos, prefixesNom);
+          if (res === undefined) return;
+          bonusCompetence = res.total;
+          found =  res.bestFit;
+        });
+      }
+      if (bonusCompetence === undefined) {
+        options.bonusAttrs = options.bonusAttrs || [];
+        options.bonusAttrs.push(options.competence);
+      } else {
+        switch (comp) {
+          case 'perception':
+            if (attributeAsBool(personnage, 'foretVivanteEnnemie')) {
+              expliquer("Forêt hostile : -5 en perception");
+              bonus -= 5;
+            }
+            break;
+          case 'survie':
+            if (attributeAsBool(personnage, 'foretVivanteEnnemie')) {
+              expliquer("Forêt hostile : -5 en survie");
+              bonus -= 5;
+            }
+            break;
+          case 'orientation':
+            if (attributeAsBool(personnage, 'foretVivanteEnnemie')) {
+              expliquer("Forêt hostile : -5 en orientation");
+              bonus -= 5;
+            }
+            break;
+          case 'discrétion':
+          case 'discretion':
+            if (attributeAsBool(personnage, 'foretVivanteEnnemie')) {
+              expliquer("Forêt hostile : -5 en discrétion");
+              bonus -= 5;
+            }
+            break;
+        }
+        var msgComp = "Compétence " + options.competence + " : ";
+        if (bonusCompetence === 0) {
+          msgComp += 0;
+        } else {
+          bonus += bonusCompetence;
+          if (bonusCompetence > 0) msgComp += "+" + bonusCompetence;
+          else msgComp += bonusCompetence;
+        }
+        expliquer(msgComp);
+      }
+    }
+    if (bonusCompetence === undefined) {
+      if (carac == 'DEX') {
+        var malusArmure = 0;
+        if (ficheAttributeAsInt(personnage, 'DEFARMUREON', 1))
+          malusArmure += ficheAttributeAsInt(personnage, 'DEFARMUREMALUS', 0);
+        if (ficheAttributeAsInt(personnage, 'DEFBOUCLIERON', 1))
+          malusArmure += ficheAttributeAsInt(personnage, 'DEFBOUCLIERMALUS', 0);
+        if (malusArmure > 0) {
+          expliquer("Armure : -" + malusArmure + " au jet de DEX");
+          bonus -= malusArmure;
+        }
+      }
     }
     // Puis la partie commune
     options = options || {};
@@ -4161,8 +4301,13 @@ var COFantasy = COFantasy || function() {
             error("Il manque un argument à l'option " + args[0], opts);
             return;
           }
+          if (options.nom && options.nom != args[1]) {
+            error("Nom du jet défini deux fois !", options.nom);
+          }
           options.nom = args[1];
-          options.bonusAttrs.push(args[1].toLowerCase());
+          if (options.competence)
+            options.bonusAttrs.push(args[1].toLowerCase());
+          else options.competence = args[1];
           return;
         case "attribut":
           if (args.length < 2) {
@@ -4272,7 +4417,6 @@ var COFantasy = COFantasy || function() {
       titre += nomJet;
       titre += "</b>";
       if (nj == 'perception') {
-        options.bonusAttrs = options.bonusAttrs || [];
         options.bonusAttrs.push('diversionManoeuvreValeur');
       }
       if (options.bonus)
@@ -8477,10 +8621,10 @@ var COFantasy = COFantasy || function() {
         var rollId = 'disparition_' + cible.token.id;
         var options1 = {...options
         };
-        options1.bonusAttrs = ["discrétion"];
+        options1.competence = "discrétion";
         var options2 = {...options
         };
-        options2.bonusAttrs = ["perception"];
+        options2.competence = "perception";
         testOppose(rollId, attaquant, "DEX", options1, cible, "SAG", options2,
           cible.messages, evt,
           function(resultat, crit, rt1, rt2) {
@@ -14602,8 +14746,8 @@ var COFantasy = COFantasy || function() {
               var pc = pointsDeChance(perso);
               if (!tr.echecCritique && pc > 0) {
                 msgRate += '<br/>' +
-                    boutonSimple("!cof-bouton-chance " + evt.id + " " + testId, "Chance") +
-                    " (reste " + pc + " PC)";
+                  boutonSimple("!cof-bouton-chance " + evt.id + " " + testId, "Chance") +
+                  " (reste " + pc + " PC)";
               }
               sendChar(charId, msgRate);
             }
@@ -15774,7 +15918,7 @@ var COFantasy = COFantasy || function() {
       }
     };
     addEvent(evt);
-    var bonusAttrs = ['vigilance', 'perception'];
+    var bonusAttrs = [];
     if (!options.nonVivant) bonusAttrs.push('radarMental');
     var display;
     if (testSurprise === undefined) {
@@ -15816,6 +15960,7 @@ var COFantasy = COFantasy || function() {
         };
         optionsTest.bonus = bonusSurprise;
         optionsTest.bonusAttrs = bonusAttrs;
+        optionsTest.competence = 'vigilance';
         testCaracteristique(perso, 'SAG', testSurprise, testId, optionsTest, evt,
           function(tr, explications) {
             var result;
@@ -23771,7 +23916,7 @@ var COFantasy = COFantasy || function() {
       if (options === undefined) return;
       if (options.son) playSound(options.son);
       var persos = [];
-      iterSelected(selection, function (perso) {
+      iterSelected(selection, function(perso) {
         persos.push(perso);
       });
       doRageDuBerserk(persos, options);
@@ -23788,7 +23933,7 @@ var COFantasy = COFantasy || function() {
     };
     var typeRage = 'rage';
     addEvent(evt);
-    persos.forEach(function (perso) {
+    persos.forEach(function(perso) {
       var attrRage = tokenAttribute(perso, 'rageDuBerserk');
       if (attrRage.length > 0) {
         attrRage = attrRage[0];
@@ -23809,8 +23954,8 @@ var COFantasy = COFantasy || function() {
               var pc = pointsDeChance(perso);
               if (!tr.echecCritique && pc > 0) {
                 msgRate += '<br/>' +
-                    boutonSimple("!cof-bouton-chance " + evt.id + " " + testId, "Chance") +
-                    " (reste " + pc + " PC)";
+                  boutonSimple("!cof-bouton-chance " + evt.id + " " + testId, "Chance") +
+                  " (reste " + pc + " PC)";
               }
               if (stateCOF.combat && attributeAsBool(perso, 'petitVeinard')) {
                 msgRate += '<br/>' + boutonSimple("!cof-bouton-petit-veinard " + evt.id + " " + testId, "Petit veinard");
@@ -29507,15 +29652,15 @@ var COFantasy = COFantasy || function() {
             } else {
               res += " => raté";
               if (stateCOF.combat &&
-                  attributeAsBool(target, 'runeForgesort_énergie') &&
-                  attributeAsInt(target, 'limiteParCombat_runeForgesort_énergie', 1) > 0) {
-                res += "</br>" + boutonSimple("!cof-bouton-rune-energie " + evt.id + " " + saveId, "Rune d'énergie");
+                attributeAsBool(perso, 'runeForgesort_énergie') &&
+                attributeAsInt(perso, 'limiteParCombat_runeForgesort_énergie', 1) > 0) {
+                res += "</br>" + boutonSimple("!cof-bouton-rune-energie " + evt.id + " " + testId, "Rune d'énergie");
               }
-              if (!tr.echecCritique) {
-                var pcTarget = pointsDeChance(target);
+              if (!testRes.echecCritique) {
+                var pcTarget = pointsDeChance(perso);
                 if (pcTarget > 0)
                   res += "</br>" + boutonSimple("!cof-bouton-chance " + evt.id + " " +
-                      saveId, "Chance") + " (reste " + pcTarget + " PC)";
+                    testId, "Chance") + " (reste " + pcTarget + " PC)";
               }
               sendChar(veCharId, res);
             }
@@ -30725,7 +30870,7 @@ on("destroy:handout", function(prev) {
 });
 
 on('ready', function() {
-  var scriptVersion = '2.15';
+  var scriptVersion = '2.16';
   on('add:token', COFantasy.addToken);
   on("change:graphic:statusmarkers", COFantasy.changeMarker);
   on("change:campaign:playerpageid", COFantasy.initAllMarkers);
@@ -31295,6 +31440,203 @@ on('ready', function() {
       a.remove();
     });
     log("Déplacement des attributs de consommables de PNJs vers la fiche");
+  }
+  if (state.COFantasy.version < 2.16) {
+    attrs = findObjs({
+      _type: 'attribute',
+    });
+      var handouts = findObjs({
+        _type: 'handout'
+      });
+    var handhoutComp = handouts.find(function(h) {
+      var handName = h.get('name');
+    return (handName == 'Compétences' || handName == 'Competences');
+    });
+    if (handhoutComp) {
+      var listeCompetences = {
+        FOR: [],
+        DEX: [],
+        CON: [],
+        SAG: [],
+        INT: [],
+        CHA: []
+      };
+      handhoutComp.get('notes', function(note) { // asynchronous
+        var carac; //La carac dont on spécifie les compétences actuellement
+    note = note.trim();
+    if (note.startsWith('<p>')) note = note.substring(3);
+    note = note.trim().replace(/<span[^>]*>|<\/span>/g, '');
+    note = note.replace(/<p>/g, '<br>');
+    note = note.replace(/<\/p>/g, '');
+    var lignes = note.trim().split('<br>');
+        lignes.forEach(function(ligne) {
+          ligne = ligne.trim();
+          var header = ligne.split(':');
+          if (header.length > 1) {
+            var c = header.shift().trim().toUpperCase();
+            if (c != 'FOR' && c != 'CON' && c != 'DEX' && c != 'INT' && c != 'SAG' && c != 'CHA') return;
+            carac = c;
+            ligne = header.join(':').trim();
+          }
+          if (ligne.length === 0) return;
+          if (carac === undefined) {
+            log("Compétences sans caractéristique associée");
+            return;
+          }
+          var comps = ligne.split(/, |\/| /);
+          comps.forEach(function(comp) {
+            if (comp.length === 0) return;
+            listeCompetences[carac].push(comp);
+          });
+        });
+      var compToCarac = {};
+      listeCompetences.FOR.forEach(function (c) {
+        compToCarac[c] = 'FOR';
+      });
+      listeCompetences.CON.forEach(function (c) {
+        compToCarac[c] = 'CON';
+      });
+      listeCompetences.DEX.forEach(function (c) {
+        compToCarac[c] = 'DEX';
+      });
+      listeCompetences.INT.forEach(function (c) {
+        compToCarac[c] = 'INT';
+      });
+      listeCompetences.SAG.forEach(function (c) {
+        compToCarac[c] = 'SAG';
+      });
+      listeCompetences.CHA.forEach(function (c) {
+        compToCarac[c] = 'CHA';
+      });
+    attrs.forEach(function (a) {
+      var attrName = a.get('name');
+      switch(attrName) {
+        case 'RACE':
+          a.set('name', 'race');
+          return;
+        case 'PROFIL':
+          a.set('name', 'profil');
+          return;
+        case 'NIVEAU':
+          a.set('name', 'niveau');
+          return;
+        case 'SEXE':
+          a.set('name', 'sexe');
+          return;
+        case 'AGE':
+          a.set('name', 'age');
+          return;
+        case 'TAILLE':
+          a.set('name', 'taille');
+          return;
+        case 'POIDS':
+          a.set('name', 'poids');
+          return;
+        case 'FORCE':
+          a.set('name', 'force');
+          return;
+        case 'DEXTERITE':
+          a.set('name', 'dexterite');
+          return;
+        case 'CONSTITUTION':
+          a.set('name', 'constitution');
+          return;
+        case 'INTELLIGENCE':
+          a.set('name', 'intelligence');
+          return;
+        case 'SAGESSE':
+          a.set('name', 'sagesse');
+          return;
+        case 'CHARISME':
+          a.set('name', 'charisme');
+          return;
+      }
+      //Les compétences
+      var charId = a.get('characterid');
+      //On ne bouge les compétences que pour les persos de type PJ
+      var typePerso = findObjs({
+        _type: 'attribute',
+        _characterid: charId,
+        name: 'type_personnage',
+      }, {
+        caseInsensitive: true
+      });
+      if (typePerso.length > 0 && typePerso[0].get('current') != 'PJ') return;
+      if (compToCarac[attrName] === undefined) return;
+      var prefix = 'repeating_competences_' + generateRowID() + '_comp_';
+      var attrSpec = {characterid:charId};
+        attrSpec.name = prefix+'nom';
+        attrSpec.current = attrName;
+        createObj('attribute', attrSpec);
+        attrSpec.name = prefix+'bonus';
+        attrSpec.current = a.get('current');
+        var attrBonus = createObj('attribute', attrSpec);
+        attrSpec.name = prefix+'bonusTotal';
+        createObj('attribute', attrSpec);
+        attrSpec.name = prefix+'carac';
+        attrSpec.current = compToCarac[attrName];
+        createObj('attribute', attrSpec);
+        if (attrSpec.current == 'DEX') {
+        attrSpec.name = prefix+'malus';
+        attrSpec.current = 'armure';
+        createObj('attribute', attrSpec);
+          attrBonus.setWithWorker('current', a.get('current'));
+        } else if (attrName == 'perception' || attrName == 'vigilance') {
+        attrSpec.name = prefix+'malus';
+        attrSpec.current = 'casque';
+        createObj('attribute', attrSpec);
+          attrBonus.setWithWorker('current', a.get('current'));
+        }
+        a.remove();
+    });
+      }); //end hand.get(notes)
+    } else {
+    attrs.forEach(function (a) {
+      var attrName = a.get('name');
+      switch(attrName) {
+        case 'RACE':
+          a.set('name', 'race');
+          return;
+        case 'PROFIL':
+          a.set('name', 'profil');
+          return;
+        case 'NIVEAU':
+          a.set('name', 'niveau');
+          return;
+        case 'SEXE':
+          a.set('name', 'sexe');
+          return;
+        case 'AGE':
+          a.set('name', 'age');
+          return;
+        case 'TAILLE':
+          a.set('name', 'taille');
+          return;
+        case 'POIDS':
+          a.set('name', 'poids');
+          return;
+        case 'FORCE':
+          a.set('name', 'force');
+          return;
+        case 'DEXTERITE':
+          a.set('name', 'dexterite');
+          return;
+        case 'CONSTITUTION':
+          a.set('name', 'constitution');
+          return;
+        case 'INTELLIGENCE':
+          a.set('name', 'intelligence');
+          return;
+        case 'SAGESSE':
+          a.set('name', 'sagesse');
+          return;
+        case 'CHARISME':
+          a.set('name', 'charisme');
+          return;
+      }
+    });
+    }
+    log("Mise à jour des attributs de compétence effectué");
   }
   state.COFantasy.version = scriptVersion;
   if (state.COFantasy.options.affichage.val.fiche.val) {